Extract CP850 to its own file
